sync propagation change, endian addition
sync propagtion through the pfb_fir was the CASPER standard of
(n_taps-1)*fft_bins but this has been changed to the Astron
standard of 0 which seems to align with theory.

Addition of parameter to allow control over endianness. This is partly
to support the existing testbench, but it might also be useful
generally.