requant_remove_lsb/msb added to pfb_fir_pipeline
requant_remove_lsb and requant_remove_msb options added to pipeline
struct that can be passed in. This is passed to the requantisation logic
after the adder trees and controls scaling to compensate for potential
signal gain (depending on coefficients).

This was hard-coded, and is now configurable (as for the Astron
version).